refactor: quiz 관련 컴포넌트들 features 디렉토리로 이동 및 네이밍 정리
