+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Home from '@/pages/Home';
-import AllComponentsTest from '@/pages/test/AllComponentTest';
-import StartQuizPage from '@/pages/quiz/StartQuizPage';
+import AllComponentsTest from './pages/test/AllComponentTest';
+import StartQuizPage from '@/pages/quiz/QuizStartPage';
 import QuizPage from '@/pages/quiz/QuizPage';
 
 function App() {
