+++ src/features/quiz/quizResult/QuizResult.types.ts
