+++ src/features/quiz/quizResult/QuizResultBlock.tsx
