+++ src/features/quiz/quizResult/QuizResultBlock.types.ts
